+++ project/src/components/login-form/login-form.tsx
@@ -61,7 +61,7 @@ export default function LoginForm(): JSX.Element {
   };
 
   return (
-    <section className="login">
+    <section className="login" data-testid="login-form">
       <h1 className="login__title">Sign in</h1>
 
       <form
